+++ backend/server.js
@@ -4,7 +4,7 @@ const { connectMongoDB } = require('./config/mongodb')
 const userRoutes = require('./routes/userRoutes')
 const authRoutes = require('./routes/authRoutes')
 const categoriesRoutes = require('./routes/categoriesRoutes')
-const purchaseRoutes = require('./routes/purchasesRoutes')
+const purchaseRoutes = require('./routes/purchaseRoutes')
 const planRoutes = require('./routes/planRoutes')
 require('colors')
 require('dotenv').config()
